Refactor PdfSection component to remove unnecessary padding and update className assignment
